Rename files from cookiecutter defaults
